+++ src/app/[lng]/(common)/notice/[id]/NoticeInfo.tsx
@@ -2,11 +2,8 @@
 
 import dayjs from 'dayjs';
 import Image from 'next/image';
-import { Trans } from 'react-i18next/TransWithoutContext';
 
 import { NoticeDetail } from '@/api/notice/notice';
-import DDay from '@/app/components/molecules/DDay';
-import Tag from '@/app/components/molecules/Tag';
 import Tags from '@/app/components/organisms/Tags';
 import { createTranslation, PropsWithLng, PropsWithT } from '@/app/i18next';
 import DefaultProfile from '@/assets/default-profile.jpeg';
